Double programming speeds for FTDI bitbang by implenting "write-only" commands
This is pretty quickly thrown together, and has the base assumption
that any single-bit twiddling does an implicit flush, because that's
what the FTDI bitbang has to do anyway. There are some extra flushes
thrown in where needed, but they may not be enough.

This brings program/verify to 3/7 seconds for a ~4500 byte binary.